Wed, Aug 17, 2022, 10:39 AM +00:00

+++ submits/readability/readability.c
@@ -10,12 +10,12 @@ int count_sentences(string text);
 int main(void)
 {
     string text1 = get_string("Write text sample: ");
-    int L = (count_letters(text1) / count_words(text1)) * 100;
-    int S = (count_sentences(text1) / count_words(text1)) * 100;
-    int index = (0.0588 * L) - (0.296 * S) - 15.8;
-    printf("%i, %i\n", L, count_letters(text1));
-    printf("%i\n", S);
-    printf("%i\n", index);
+    float L = (count_letters(text1) / count_words(text1)) * 100;
+    float S = (count_sentences(text1) / count_words(text1)) * 100;
+    float index = (0.0588 * L) - (0.296 * S) - 15.8;
+    printf("%f, %i\n", L, count_letters(text1));
+    printf("%f\n", S);
+    printf("%f\n", index);
     if (index > 16)
     {
         printf("Grade 16+\n");
@@ -26,7 +26,7 @@ int main(void)
     }
     else
     {
-        printf("Grade %i\n", index);
+        printf("Grade %f\n", index);
     }
 }
